changed error density introduced a new hashing function (WIP) converted time offsets to show upto 1 decimal point

+++ src/AnalyzableSamples.java
@@ -86,7 +86,7 @@ public abstract class AnalyzableSamples {
      * @param aS2
      * @return
      */
-    public abstract int[] getMatchPositionInSeconds(AnalyzableSamples aS2);
+    public abstract double[] getMatchPositionInSeconds(AnalyzableSamples aS2);
     
 
     /**
